Move pl011 emulation to own header
